requirements.txt
discord.py>=2.4.0,<3.0.0
APScheduler>=3.10.4,<4.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.31.0,<3.0.0
google-api-python-client>=2.160.0,<3.0.0
google-auth-oauthlib>=1.2.1,<2.0.0
google-auth-httplib2>=0.2.0,<1.0.0
tzdata>=2024.1
